+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Update `ControlRoot` and `ControlMember` to be structs instead of newtypes. This allows eliding the container when using anonymous ids.
 - Remove `Copy` derive from `AnimationSettings`, `AnimationConfig`, and `LoopedAnimationConfig` to avoid silent copies causing issues when editing values.
 - Expand API of `NodeAttributes` component to make editing attributes easier.
+- Validate loadable names on registration. Only named structs that start uppercase are allowed.
 
 ## 0.5.1